backend/app/requirements.txt
cryptography
Flask
Flask-MySQL
PyMySQL
flask-cors
mysql-connector-python
speechrecognition